major upgrade

+++ utils/cleanup/cleanup_dir_trees.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# spellcheck source=./utils/helpers/validation.sh
+source ./utils/helpers/validation.sh
+
 # export ROOT dir name to .bashrc so it canbe referenced and deleted
 # uninstall one or more installed programs
 #   - present list of those that are installed by using same pre-install checks
@@ -18,9 +21,15 @@
 # IFS=" " read -r -a array <<< $(grep ROOT_ENV_DIR* ~/.profile)
 
 # convert grep search output with ROOT_ENV_DIR prefix to array
-arr=($(grep "ROOT_ENV_DIR" ~/.profile))
+arr=()
+if [[ -f ~/.profile ]] ; then
+  arr=($(grep "ROOT_ENV_DIR" ~/.profile))
+fi
 # remove entries in array that are just "export"
 no_exports_arr=( "${arr[@]/export/}" )
+# EX: ROOT_ENV_DIR_NAME=/c/Users/<user>/<ROOT_ENV_DIR> - Windows
+# EX: ROOT_ENV_DIR_NAME=/home/<user>/<ROOT_ENV_DIR> - Linux
+# EX: ROOT_ENV_DIR_NAME=/Users/<user>/<ROOT_ENV_DIR> - Darwin
 # flatten empty spaces out of array
 ENV_DIR_ARRAY=(${no_exports_arr[@]})
 # get array length
@@ -92,28 +101,49 @@ iteratively_remove_env_dirs() {
   echo -n "> "
   read -r indexes
 
-  if [[ ${indexes} == 'all' ]] ; then
-    remove_all_dirs
-  else
-    # check_and_install_sed
-    for i in ${indexes[@]}
-    do
-      # IS_ENV_VAR_IN_PROFILE="$(grep -wnq ${!ENV_DIR_NAMES[i]} ~/.profile)"
-      # w: strict word search, n: output with line number
-      ENV_VAR_IN_PROFILE="$(grep -wn ${ENV_DIR_NAMES[i]} ~/.profile)"
-      # split by colon to get line number
-      IFS=":" read -r -a arr <<< ${ENV_VAR_IN_PROFILE}
+  while "true"
+  do
+    if [[ "$(input_is_number_with_possible_spaces "${indexes}")" == "true" ]] || [[ "$(input_is_the_word_all "${indexes}")" == "true" ]]; then
+      if [[ ${indexes} == 'all' ]] ; then
+        remove_all_dirs
+      else
+        # check_and_install_sed
+        for i in ${indexes[@]}
+        do
+          # IS_ENV_VAR_IN_PROFILE="$(grep -wnq ${!ENV_DIR_NAMES[i]} ~/.profile)"
+          # w: strict word search, n: output with line number
+          ENV_VAR_IN_PROFILE="$(grep -wn ${ENV_DIR_NAMES[i]} ~/.profile)"
+          # split by colon to get line number
+          IFS=":" read -r -a arr <<< ${ENV_VAR_IN_PROFILE}
 
-      # evaluates the string stored in the variable to its variable definition
-      if [[ -d "${ENV_DIR_PATHS[i]}" ]] && [[ -n ${ENV_VAR_IN_PROFILE} ]] ; then
-        # remove directory
-        rm -rf "${ENV_DIR_PATHS[i]}"
-        # remove env var by line number
-        sed -i "${arr[0]}d" ~/.profile
-      elif [[ ! -d "${ENV_DIR_PATHS[i]}" ]] && [[ -n ${ENV_VAR_IN_PROFILE} ]] ; then
-        # remove env var by line number
-        sed -i "${arr[0]}d" ~/.profile
+          # evaluates the string stored in the variable to its variable definition
+          if [[ -d "${ENV_DIR_PATHS[i]}" ]] && [[ -n ${ENV_VAR_IN_PROFILE} ]] ; then
+            echo
+            echo "==================================================="
+            echo "Removing: ${ENV_DIR_PATHS[i]}..."
+            # remove directory
+            rm -rf "${ENV_DIR_PATHS[i]}"
+            echo "Removing exported variable..."
+            # remove env var by line number
+            sed -i "${arr[0]}d" ~/.profile
+            echo "==================================================="
+          elif [[ ! -d "${ENV_DIR_PATHS[i]}" ]] && [[ -n ${ENV_VAR_IN_PROFILE} ]] ; then
+            echo
+            echo "==================================================="
+            echo "Removing only exported variable for ${ENV_DIR_PATHS[i]}"
+            # remove env var by line number
+            sed -i "${arr[0]}d" ~/.profile
+            echo "==================================================="
+          fi
+        done
       fi
-    done
-  fi
+      break
+    else
+      echo
+      echo "Input must be a number in the above list, or 'all'"
+      echo
+      echo -n "> "
+      read -r indexes
+    fi
+  done
 }
